Higher level function for using bootstrap_SRS with SGP objects

Function for calculating bootstrap summary statistics using data and pre-designated strata proportions with SGP objects.

bootstrapSRS_SGP(
        sgp_object,
        strata_summaries=c("STATUS", "GROWTH"),
        strata_variables="SCALE_SCORE_DECILE",
        strata_proportions_years_status,
        strata_proportions_years_growth,
        summary_years,
        create_scale_score_deciles=TRUE,
        sample_size=NULL,
        bootstrap_iterations=100,
        summary_statistic="mean",
        aggregation_group=c("CONTENT_AREA", "GRADE"))

Arguments

sgp_object

strata_summaries
Character string indicating STATUS and/or GROWTH.

strata_variables

strata_proportions_years_status

strata_proportions_years_growth

summary_years

create_scale_score_deciles

sample_size

bootstrap_iterations

summary_statistic

aggregation_group

Details

Function calculates a summary statistic of existing data (e.g., mean) by re-sampling based upon designated strata proportions

Value

Function returns either a list (default) containing the summary statistic (e.g., mean) and standard deviation (i.e., standard error) for the bootstrap sample of that statistic or funtion returns a vector containing the bootstrap replication values.

Examples

### bootstrapSRS_SGP needs an object of class SGP run.
### See the Step 6 comparisons in the Demonstration Learning Loss Analytics for examples
### https://github.com/CenterForAssessment/SGP_Research/tree/master/Demonstration/Learning_Loss_Analysis/Step_6_Summary_Comparisons