Diagnostic achievement gap information

Function for creating diganostic plots depicting achievement and other types of gaps derived from education assessment data.

cfaGaps(
        gap_data,
        gap_groups=NULL,
        gap_plot_title="Non-Parametric Gap Plot",
        gap_plot_fig_width=13,
        gap_plot_fig_height=8.5,
        gap_plot_file_path="Visualizations/Gaps/",
        gap_plot_file_name="Gaps",
        gap_plot_output_format="PDF")

Arguments

gap_data
Assessment data (usually scaled scores) from which to calculate and depict (achievement) gaps.

gap_groups
Vector of group names with which to calculate/create gap diagnostic information/plots.

gap_plot_title
Character string indicating the title of the gap plot. Default is 'Non-paerametric gap plot'.

gap_plot_fig_width
Width (in inches) of gap plot. Default is 13 inches.

gap_plot_fig_height
Height (in inches) of gap plot. Default is 8.5 inches.

gap_plot_file_path
Path to be used where gap plot is saved. Default is 'Visualizations/Gaps/'.

gap_plot_file_name
File name to be used for gap plot. Default is 'Gaps'.

gap_plot_output_format
Character string indicating the output format for the figures produced by cfaGaps. Currently supports PDF, PNG, SVG, and PLOTLY.

Details

Typical use of the function is to submit scale score data together with group identifiers from which to calculate and depict (achievement) gaps.

Value

Function returns plots depicting/representing (achievement) gap data).

Examples

        not_run({
                cfaGaps(gap_data=sgpData_LONG[YEAR=="2018_2019", c("SCALE_SCORE", "ETHNICITY"), with=FALSE],
                                                gap_groups=c("African American", "White"))

})